+ CS599 Machine Learning Programming assignment 03 completed

+++ Project3_Kmean/CS599_u.nair_PA3/CS599_u.nair_PA3/positionVector.h
@@ -3,8 +3,8 @@
 class positionVector
 {
 public:
-	int xposition;
-	int yposition;
+	float xposition;
+	float yposition;
 	positionVector();
 	~positionVector();
 };
